port to vscode mplab ide, start on tile streaming

+++ moduleDogm128.c
@@ -217,7 +217,7 @@ void drawPixel(uint8_t x, uint8_t y, uint8_t color)
         displayBuffer[page][x] &= ~(1 << bit);
 }
 
-void drawLine(int x0, int y0, int x1, int y1)
+void drawLine(uint8_t x0, uint8_t y0, uint8_t x1, uint8_t y1)
 {
     int dx = abs(x1 - x0);
     int sx = x0 < x1 ? 1 : -1;
